Move Caddy module to `/outlinecaddy`.
