+++ outlinecaddy/cmd/caddy/main.go
@@ -24,7 +24,7 @@ package main
 import (
 	caddycmd "github.com/caddyserver/caddy/v2/cmd"
 
-	_ "github.com/Jigsaw-Code/outline-ss-server/caddy"
+	_ "github.com/Jigsaw-Code/outline-ss-server/outlinecaddy"
 	_ "github.com/caddyserver/caddy/v2/modules/standard"
 	_ "github.com/iamd3vil/caddy_yaml_adapter"
 	_ "github.com/mholt/caddy-l4"
