+++ outlinecaddy/examples/websocket.yml
